rcdzc: extract backend-agnostic Core-IR analysis primitives to a shared core_analysis module
Pure move (no logic change): core_eq, core_hash_key, collect_node_refs,
collect_dominating_frontier, subtree_size, is_heap_type, licm_children — all
verified zero backend/slot/Lir state — relocate from backend/wasm/select.rs into
a new crate-level core_analysis module (pub(crate)). select.rs re-imports them so
its Lir-level LICM/CSE call sites are unchanged. This lets the backend-independent
Core optimization passes (opt.rs) reuse the exact same soundness-critical frontier
+ heap-type logic without duplicating it (two copies would risk drifting on the
guards that keep an LICM/CSE hoist sound). The realization-coupled predicates stay
in select.rs — is_cse_shareable bakes in the wasm CSE hoist-before-body assumption
(excludes Core::LocalRef, unbound slot at the hoist point), which a Core pass that
introduces the let does not share, so it is deliberately not extracted.

Coordinated with v-core-opt (building a Core O2 CSE pass on this shared module).
Behavior-identical: rcdzc lib 2439/0, tests::recursion 33/0, fmt+clippy+build clean.
